CI: Cordova - Build iOS
This implements building the iOS package + uploading the artifact.

It builds successfully though I was unable to get past the loading screen, even locally (not sure what's wrong, might need help).

The QRScanner problem was a journey. It started with trying to install the package via url, from https://github.com/padloc/cordova-plugin-qrscanner and I ran into https://github.com/lerna/lerna/issues/2674 then I tried to simply publish a new package name (`padloc-cordova-plugin-qrscanner`), but as I was replacing the plugin name in the code, it started blowing up in a few different places, and it just felt wrong. Finally, I just created a patch command in `npm`, which can later evolve into a bash script or whatever we need. It's not pretty, but that worked!

The signed command doesn't run yet for iOS because building for the app store will be a separate PR, for publishing, like the other targets.

This will close the development/CI/CD part of #330