Find Item e2e test
This adds a "find item" e2e test that confirms matching items show, and non-matching items don't.

I also implemented the `waitForAnimations` config, but I found the results inconsistent via CLI, so I left those in, after much experimentation.

There are a couple of improvements on reducing the number of `.find()` calls in the Cypress tests, when the elements aren't rendered in the Shadow DOM.