Enable Auto-lock in web extension
This allows the web extension to auto-lock (tested in Chrome and Firefox), by running an alarm in the background, since the `setTimeout` is ignored past 10 seconds.

I noticed that if the popup is open, it does not lock until it’s closed (even after firing the alarm). I couldn’t find a way to change the route for the app from the background, and I wasn’t sure if it would worth it adding a new listener for this, as it’s not common for the extension to be open for a long period of time anyway.

Closes #300