Enable SAML SSO for SCIM (Google Workspaces)
This adds support for integrating SCIM with Google Workspaces (or Cloud Identity) via SAML SSO.

Basically, when someone initiates SSO from their provider, it'll go to the right URL with the authorization info (once authorized in the provider), and we'll take that information, transform to a SCIM (new user) request, and parse that in.

It's not very pretty, but it works. This needs to be tested and still needs some polish (redirect to the start page once the user has been created or already exists).