Desktop lock on suspend
Related to #211

Note that this attempt doesn't quite work as the app doesn't load properly in Electron, and I eventually stopped trying to, since it's turning into a lot of effort. I just wanted it "documented".

There are to options to close #211:

1. When we detect the system is suspending/locking, shutdown/quit the app. This won't require any communication with the PWA.
2. We implement something like [`electron-safe-ipc`](https://www.npmjs.com/package/electron-safe-ipc) directly in the PWA, so electron can send the lock event to it.

Option 1 is not a great UX, but does handle the security aspect.

Option 2 is a much better UX, but introduces a new dependency that is only _relevant_ when the app's used via Electron.