Fix changing master password + UX/UI improvements
This fixes changing the master password and adds some UI info for better UX to the device/session list.

Basically the problem was that we were trying to unlock/decrypt the data with the new password, before encrypting it with it (so it was encrypted with the old password still). Now we decrypt with the old password and encrypt with the new one.

I've tested this with shared vaults as well.

Fixes #527
Fixes #516