Make release assets consistent
Unfortunately [it's not yet possible for Tauri](https://github.com/tauri-apps/tauri-action/issues/215), nor there's any GH action for that, _but_ [there's an API endpoint](https://docs.github.com/en/rest/releases/assets#update-a-release-asset) we could use.

Since it's only the Tauri assets that are not according to the new standard, I think it's fine to keep it something we manually update, otherwise we can try to build a script that taps into that endpoint, later.