Finish adding auditor logs for anything except items
I built an API endpoint for logging changes to items individually, but as I was working on the encryption for the item data before sending it to the server, I wasn't happy with how that could be abused (even though it required authentication), so I ditched it, and decided to instead look through and log item changes when a vault is updated.

While it might technically be slower (we'll waste a lot of time looping through items that weren't changed), it'll be more accurate.

Unfortunately I only commented where that will happen, I haven't built that yet.