Rework the audit log system to remove data storage and send item logs from the client.
Note that, for some reason I couldn't quite figure out (hopefully it'll be quick and clear for @MaKleSoft), the params sent to the `itemAuditorLog` endpoint are an empty object (they're OK on the frontend, just not deserialized properly in the backend).

I've also tested the other parts with a postgresql backend, with success.