View and restore history entries, for all relevant fields, with a nice UI/UX.
I started implementing the delete option, but it wasn't looking very clean (passing around a `historyEntryIndex` to `updateItem`), so I ditched it for now, as you can simply make small edits to "make it disappear".

Right now we're only missing the `vaultId` change (logging and restoring), but as I was thinking about implementing its restore, I'm also wondering if we need it, since restoring to a previous vault might not really work (it might no longer exist, or you may no longer have access), and it seems different than the other changes we're logging and restoring to, here.